requirements.txt
alpaca-py>=0.21.0
lumibot>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
ta>=0.11.0
loguru>=0.7.0
schedule>=1.2.0
backtesting>=0.3.3
typer[all]>=0.9.0
httpx>=0.24.0
rich>=13.0.0
